102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -4,16 +4,16 @@
  * main - Entry point of the program.
  *
  * Description: This function generates and prints the first 50 Fibonacci
- *              numbers separated by commas and a space, starting with 1 and 2.
+ * numbers separated by commas and a space, starting with 1 and 2.
  *
  * Return: Always 0 (success).
  */
 int main(void)
 {
-	long long int a = 1, b = 2, temp;
+	long a = 1, b = 2, temp;
 	int count = 0;
 
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 	count += 2;
 
 	while (count < 50)
@@ -23,9 +23,9 @@ int main(void)
 		b = temp;
 
 		if (count == 49)
-			printf("%lld\n", temp);
+			printf("%ld\n", temp);
 		else
-			printf("%lld, ", temp);
+			printf("%ld, ", temp);
 
 		count++;
 	}
